chore: id type Long에서 Integer 로 변경

+++ src/main/java/com/idt/aio/repository/DocumentImageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DocumentImageRepository extends JpaRepository<DocumentImage, Long> {
+public interface DocumentImageRepository extends JpaRepository<DocumentImage, Integer> {
 }
